refact: article package 내부 component 분리
