+++ app/src/main/java/com/dev/briefing/presentation/detail/ArticleScreen.kt
@@ -1,11 +1,7 @@
 package com.dev.briefing.presentation.detail
 
-import android.app.Activity
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
 import android.util.Log
-import android.widget.Toast
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
@@ -14,7 +10,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.ColorFilter
@@ -24,25 +19,16 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.core.content.ContextCompat
 import com.dev.briefing.R
-import com.dev.briefing.data.model.Article
 import com.dev.briefing.data.model.BriefingDetailResponse
 import com.dev.briefing.data.model.tmpBriefingResponse
 import com.dev.briefing.presentation.theme.GradientEnd
 import com.dev.briefing.presentation.theme.GradientStart
 import com.dev.briefing.presentation.theme.MainPrimary
-import com.dev.briefing.presentation.theme.MainPrimary4
 import com.dev.briefing.presentation.theme.SubText2
 import com.dev.briefing.presentation.theme.Typography
 import com.dev.briefing.presentation.theme.White
-import com.dev.briefing.presentation.theme.utils.CommonDialog
-import com.dev.briefing.util.MEMBER_ID
-import com.dev.briefing.util.MainApplication
 import com.dev.briefing.util.SCRAP_TAG
 import org.koin.androidx.compose.getViewModel
 import org.koin.core.parameter.parametersOf
@@ -98,108 +84,6 @@ fun ArticleDetailScreen(
                 Spacer(modifier = Modifier.height(25.dp))
             }
         }
-    }
-}
-
-@Composable
-fun DetailHeader(
-    onBackClick: () -> Unit,
-    scrap: () -> Boolean,
-    unScrap: () -> Boolean,
-    rank: Int,
-    context: Context,
-    isScrap: Boolean
-) {
-    // 로그인이 안된 상태에서 스크랩 시도시 로그인 다이얼로그
-    val openLogInDialog = remember { mutableStateOf(false) }
-    if (openLogInDialog.value) {
-        CommonDialog(
-            onDismissRequest = { openLogInDialog.value = false },
-            onConfirmation = {
-                openLogInDialog.value = false
-                (context as Activity).finish()
-            },
-            dialogTitle = R.string.dialog_login_title,
-            dialogText = R.string.dialog_login_text,
-            dialogId = R.string.dialog_login_confirm,
-            confirmColor = MainPrimary4
-        )
-    }
-    /**  스크랩 관련 변수
-     * 1. scrap img
-     * 2.
-     */
-
-    //1. scrapImg resource 정의
-    val unScrapImg = painterResource(id = R.drawable.scrap_normal)
-    val scrapImg = painterResource(id = R.drawable.scrap_selected)
-
-    //2. scrapStatus 정의
-    val isScrapStatus = remember {
-        //파라미터로 받은 isScrap값으로 초기화 해준다
-        mutableStateOf(false)
-    }
-    LaunchedEffect(isScrap) {
-        isScrapStatus.value = isScrap
-    }
-    Log.d(SCRAP_TAG, "1. 파라미터로 전달받은 isScrap 값: ${isScrap}, isScrapStatus 값: ${isScrapStatus}")
-
-
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 60.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    )
-    {
-        Image(
-            painter = painterResource(
-                id = R.drawable.arrow
-            ),
-            contentDescription = "back Key", modifier = Modifier
-                .clickable(onClick = onBackClick)
-        )
-        Text(
-            text = "Briefing #${rank}",
-            style = Typography.titleMedium.copy(
-                color = White,
-                fontWeight = FontWeight(400)
-            )
-        )
-        Image(
-            painter = if (isScrapStatus.value) scrapImg else unScrapImg,
-            contentDescription = if (isScrapStatus.value) "Unliked" else "Liked",
-            modifier = Modifier.clickable(
-                onClick = {
-                    Log.d(SCRAP_TAG, "2. 클릭이벤트 발생! 변경전 scrap 값 : ${isScrapStatus.value}")
-                    val memberId: Int = MainApplication.prefs.getSharedPreference(MEMBER_ID, 0)
-                    if (memberId != 0) {
-                        if (isScrapStatus.value) {
-                            Log.d(SCRAP_TAG, "3. 스크랩 취소 : ${isScrapStatus.value}")
-                            if (unScrap()) {
-                                isScrapStatus.value = false
-                            }else{
-                                Toast.makeText(context, "스크랩 해제에 실패하였습니다. 다시 시도해주세요", Toast.LENGTH_SHORT).show()
-                            }
-                        } else {
-                            Log.d(SCRAP_TAG, "3. 스크랩 등록")
-                            if (scrap()) {
-                                isScrapStatus.value = true
-                            }else{
-                                Toast.makeText(context, "스크랩에 실패하였습니다. 다시 시도해주세요", Toast.LENGTH_SHORT).show()
-                            }
-                        }
-                        Log.d(SCRAP_TAG, "4. 클릭이벤트 발생! 변경전 scrap 값 : ${isScrapStatus.value}")
-
-                    } else {
-                        openLogInDialog.value = true
-                    }
-                }
-            )
-        )
-
     }
 }
 
@@ -266,9 +150,6 @@ fun ArticleDetail(
             text = stringResource(R.string.detail_article_header),
             style = Typography.headlineLarge
         )
-        if (article.isBriefOpen) {
-            BriefChatLink()
-        }
         Column(
             verticalArrangement = Arrangement.spacedBy(13.dp)
         ) {
@@ -281,91 +162,5 @@ fun ArticleDetail(
         }
         Spacer(modifier = Modifier.height(25.dp))
 
-    }
-}
-
-@Composable
-fun ArticleLink(
-    newsLink: Article,
-    modifier: Modifier = Modifier,
-    context: Context
-) {
-    Row(
-        modifier
-            .fillMaxWidth()
-            .background(White, shape = RoundedCornerShape(40.dp))
-            .border(1.dp, MainPrimary, shape = RoundedCornerShape(10.dp))
-            .clickable {
-                val intent = Intent(Intent.ACTION_VIEW, Uri.parse(newsLink.url))
-                ContextCompat.startActivity(context, intent, null)
-            }
-            .padding(vertical = 9.dp, horizontal = 13.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-
-        Column(
-            modifier = Modifier.widthIn(max = 193.dp)
-        ) {
-            Text(
-                text = newsLink.press, style = Typography.bodyMedium.copy(
-                    fontWeight = FontWeight(700),
-                    color = MainPrimary
-                )
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-            Text(
-                text = newsLink.title,
-                style = Typography.labelSmall,
-                overflow = TextOverflow.Ellipsis
-            )
-        }
-        Image(painter = painterResource(id = R.drawable.left_arrow), contentDescription = "fdfd")
-    }
-}
-
-/**
- * [ChatScreen]으로 이동하는 버튼
- */
-@Preview
-@Composable
-fun BriefChatLink(
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier
-            .fillMaxWidth()
-            .background(White, shape = RoundedCornerShape(40.dp))
-            .border(1.dp, MainPrimary, shape = RoundedCornerShape(10.dp))
-            .clickable {
-                //TODO: add ChatScreen navigate
-            }
-            .padding(vertical = 9.dp, horizontal = 13.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.chat_selelcted),
-            contentDescription = "fdfd"
-        )
-        Spacer(modifier = Modifier.width(9.dp))
-        Text(
-            text = stringResource(R.string.detail_brief_text), style = Typography.bodyMedium.copy(
-                fontWeight = FontWeight(700),
-                color = MainPrimary
-            )
-        )
-        Text(
-            text = stringResource(R.string.detail_brief_label),
-            style = Typography.labelSmall.copy(
-                color = MainPrimary,
-                fontSize = 7.sp,
-                lineHeight = 8.sp
-            ),
-            overflow = TextOverflow.Ellipsis,
-            textAlign = TextAlign.Start
-        )
-
-        Image(painter = painterResource(id = R.drawable.left_arrow), contentDescription = "fdfd")
     }
 }
